Simplify provider search filters
