+++ apps/web/utils/email/types.ts
@@ -247,6 +247,8 @@ export interface EmailProvider {
     query: string;
     maxResults?: number;
     pageToken?: string;
+    readState?: "read" | "unread";
+    labelName?: string;
   }): Promise<{
     messages: ParsedMessage[];
     nextPageToken?: string;
